// FRONTEND/pryGestionDeAgua_Cliente/src/app/Interfaces/miCuenta.ts
export interface miCuentaInterface {
    dni: string;
    nombre: string;
    apellido: string;
    mail: string;
    password: string;
    fechaNacimiento: string;
    idRol?: number;
    calle?: string;
    piso?: string;
    depto?: string;
    telefono?: string;
    idBarrio: number;
    sexo:string;
}